Using maven wrapper instead of gradle
Having used gradle at work + in this project, it's nice - its compact and can do pretty much all thigns maven can do, if not more. But it's still too new and overall feels like workarounds were made to get plugins to work, instead of having it just work.

If for some reason in the future I feel gradle is the obvious and superior option to maven, I'll switch back. But it seems to me maven is just the better choice at this point of time.